Move reportio pkg into common internal dir.
