+++ cmd/ingest/main.go
@@ -28,11 +28,11 @@ import (
 	_ "gocloud.dev/blob/gcsblob"
 	_ "gocloud.dev/blob/s3blob"
 
-	"github.com/ossf/malicious-packages/cmd/ingest/reportio"
 	"github.com/ossf/malicious-packages/cmd/ingest/sourceio"
 	"github.com/ossf/malicious-packages/cmd/ingest/startkeys"
 	"github.com/ossf/malicious-packages/internal/config"
 	"github.com/ossf/malicious-packages/internal/report"
+	"github.com/ossf/malicious-packages/internal/reportio"
 	"github.com/ossf/malicious-packages/internal/source"
 )
